backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
gdown
python-jose[cryptography]
passlib[bcrypt]
python-multipart
